[#58] Move: EmailAuth 패키지 이동
